Create ProductStockCard component, add loading placeholder to stock panel.
